限制弱密码用户使用addhome命令

+++ server/src/commands/用户/addhome.js
@@ -29,6 +29,13 @@ export async function run(core, server, socket, data) {
       text:'要注册个人空间，就必须拥有一个识别码。'
     },socket)
   }
+
+  if (socket.passwordWarning){
+    return server.reply({
+      cmd:'warn',
+      text:'抱歉，要使用该功能，您必须设置一个强密码。\n一个强密码的定义：最少6位，包括至少1个大写字母，1个小写字母，1个数字，1个特殊字符\n如果您继续使用弱密码，那么您可能无法正常使用部分功能。'
+    },socket)
+  }
   
   if (!data.text){
     core.chatDB.queryData(`delete from home where trip = '${socket.trip}';`,ret=>{
